Fix build to work with VS2010
Needed to play with the custom build tool definitions to define in
a few projects (driver and dcapi) to use yasm.  Also needed to
modify boot_stage0 project to use $ to reference environment variables
instead of %

Builds now work on both x64 and 32 bit platforms

Be careful when installing - I've noticed the x64 bit install fails on
my system after reboot (needed to run system recovery).  Be sure to
test all builds in a virtual machine before running on a development
machine